import perl 5.10.0 from CPAN

+++ gnu/usr.bin/perl/lib/Memoize/t/tie_ndbm.t
@@ -26,16 +26,14 @@ if ($@) {
   exit 0;
 }
 
+if (! -w $ENV{TMP}) {
+  print "1..0\n";
+  exit 0;
+}
+
 print "1..4\n";
 
-
-if (eval {require File::Spec::Functions}) {
- File::Spec::Functions->import();
-} else {
-  *catfile = sub { join '/', @_ };
-}
-$tmpdir = $ENV{TMP} || $ENV{TMPDIR} ||  '/tmp';  
-$file = catfile($tmpdir, "md$$");
+$file = "md$$";
 1 while unlink $file, "$file.dir", "$file.pag", "$file.db";
 tryout('Memoize::NDBM_File', $file, 1);  # Test 1..4
 1 while unlink $file, "$file.dir", "$file.pag", "$file.db";
